LayoutTests/media/modern-media-controls/ios-inline-media-controls/touch/ios-inline-media-controls-added-after-play-expected.txt

Check that tapping the start button pauses the video if controls are added after it has started playing.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


'play' event.
PASS media.paused is false
PASS shadowRoot.querySelector('button.play-pause') became different from null
PASS shadowRoot.querySelector('button.play-pause').getBoundingClientRect().width became different from 0
Tapping start button...
Waiting for pause...
PASS media.paused became true
Checking again...
PASS media.paused is true
PASS successfullyParsed is true

TEST COMPLETE
